Change directory structure and seperate web and tracker into seperate apps
Web and Tracker are now two seperate applications implemented in web/ and tracker/ directories. Models are shared between them in the models/ directory. Each application is run with the app.js file in it's directory. An app.js is placed in the root directory which runs the application specified in its arguments. Applications should be run using this since some initializations (such as creating log directories) are done in here.
The config.js in the root directory specifies overall configurations such as db settings and urls. Application specific settings such as routes and express configurations are moved to settings/ directory in each app.
Logging is not shared and has been moved to applications settings/ directory.